Fix test warning cleanup and helper imports
